requirements.txt
pygsheets==2.0.5
python-telegram-bot==13.6
pandas==1.2.4
opencv-python==4.5.2.54
numpy==1.20.3
python-dotenv==0.18.0